+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Add support for more type annotations on methods
   - Add support for modifying types reflecting `defaultProps`
   - Add support for `React.PureComponent`
+  - Add support for generics
 
 ## 0.1.5